Step 2 Code clean up
Consolidated more unused files.

main.js now only calls IRSpectrum.py (though it needs to call it twice. I don't know how to get main.js to correctly recieve output from CompareQueryToDB.py otherwise)
