+++ IRSpectrum Testing only data/query/Thumbs.db
