+++ IRSpectrum Testing only data/src/jpegExtract.py
