+++ IRSpectrum Testing only data/tempPics/90-01-7.jpg
