+++ IRSpectrum Testing only data/tempPics/Query1.jpg
