+++ IRSpectrum Testing only data/tempPics/graph.png
