+++ IRSpectrum Testing only data/utilities/black.jpg
