+++ IRSpectrum Testing only data/utilities/placeholder.png
