+++ output/Ranked Differences.txt
@@ -1,105 +1,105 @@
-#1: 110-85-0_a.pdf
+#1: 98-95-3_a.pdf
 difference = 0.0
 
-#2: 110-85-0_c.pdf
-difference = 60.56829956019463
+#2: 98-95-3_c.pdf
+difference = 54.41100998501643
 
-#3: 110-85-0_b.pdf
-difference = 89.64299511749822
+#3: 100-02-7_a.pdf
+difference = 132.59786011985184
 
-#4: 623-05-2.pdf
-difference = 105.72803775682596
+#4: 50-78-2_a.pdf
+difference = 139.42606188076843
 
-#5: 591-27-5.pdf
-difference = 115.25568123951511
+#5: 50-78-2_b.pdf
+difference = 146.2006685647666
 
-#6: 556-18-3.pdf
-difference = 120.65618166708853
+#6: 110-86-1_b.pdf
+difference = 146.73079316451594
 
-#7: 31230-17-8.pdf
-difference = 129.20296138893522
+#7: 100-02-7_b.pdf
+difference = 160.41033104661273
 
-#8: 55-21-0.pdf
-difference = 129.270677834137
+#8: 62-53-3_b.pdf
+difference = 162.93555001241856
 
-#9: 5344-90-1.pdf
-difference = 136.02065923235207
+#9: 99-05-8.pdf
+difference = 171.9593030034193
 
-#10: 103-90-2.pdf
-difference = 138.86404194211462
+#10: 62-53-3_c.pdf
+difference = 183.55721776347826
 
-#11: 16082-33-0.pdf
-difference = 149.6392628258834
+#11: 89951-56-4.pdf
+difference = 198.81055417050493
 
-#12: 89951-56-4.pdf
-difference = 151.18432004767573
+#12: 103-90-2.pdf
+difference = 206.08953702722172
 
-#13: 623-04-1.pdf
-difference = 177.49107644839629
+#13: 120-72-9_a.pdf
+difference = 211.83956435464611
 
-#14: 100-51-6_c.pdf
-difference = 177.99572479684693
+#14: 620-24-6.pdf
+difference = 231.01075633770716
 
-#15: 90-01-7.pdf
-difference = 188.6881035347818
+#15: 100-46-9_c.pdf
+difference = 234.19805989489132
 
-#16: 100-02-7_b.pdf
-difference = 198.62108995043232
+#16: 100-46-9_a.pdf
+difference = 249.31829216395084
 
-#17: 620-24-6.pdf
-difference = 199.10355407921622
+#17: 110-85-0_b.pdf
+difference = 250.8319772496008
 
-#18: 50-78-2_b.pdf
-difference = 208.7572256830621
+#18: 591-27-5.pdf
+difference = 252.1378806761845
 
-#19: 100-46-9_a.pdf
-difference = 213.59616714851998
+#19: 108-88-3_b.pdf
+difference = 253.15109479866788
 
-#20: 99-05-8.pdf
-difference = 213.90608215767722
+#20: 108-88-3_c.pdf
+difference = 255.24610630921495
 
-#21: 100-51-6_a.pdf
-difference = 221.867254830945
+#21: 90-01-7.pdf
+difference = 261.70292771488687
 
-#22: 100-46-9_c.pdf
-difference = 224.534377546884
+#22: 55-21-0.pdf
+difference = 262.153198478478
 
-#23: 120-72-9_b.pdf
-difference = 224.787542242553
+#23: 623-04-1.pdf
+difference = 277.8024166601764
 
-#24: 50-78-2_a.pdf
-difference = 224.8694715930898
+#24: 110-85-0_c.pdf
+difference = 278.0905803890612
 
-#25: 62-53-3_c.pdf
-difference = 225.481517768715
+#25: 5344-90-1.pdf
+difference = 283.5534685098453
 
-#26: 62-53-3_b.pdf
-difference = 260.1173559812914
+#26: 120-72-9_b.pdf
+difference = 285.6860372018014
 
-#27: 100-02-7_a.pdf
-difference = 262.7492363744477
+#27: 623-05-2.pdf
+difference = 292.6525612307169
 
-#28: 110-89-4_c.pdf
-difference = 277.8615877520896
+#28: 556-18-3.pdf
+difference = 311.71487326672695
 
-#29: 108-88-3_c.pdf
-difference = 278.1826108833613
-
-#30: 110-86-1_b.pdf
-difference = 291.39315243262706
-
-#31: 110-89-4_b.pdf
-difference = 308.40390605220114
-
-#32: 108-88-3_b.pdf
-difference = 313.1494051152967
-
-#33: 98-95-3_a.pdf
+#29: 110-85-0_a.pdf
 difference = 316.365692772295
 
-#34: 120-72-9_a.pdf
-difference = 352.82319160671847
+#30: 100-51-6_a.pdf
+difference = 341.6181726756536
 
-#35: 98-95-3_c.pdf
-difference = 359.8304076009568
+#31: 100-51-6_c.pdf
+difference = 362.0507645376052
+
+#32: 16082-33-0.pdf
+difference = 406.19657869197397
+
+#33: 31230-17-8.pdf
+difference = 411.4683634872034
+
+#34: 110-89-4_c.pdf
+difference = 513.1062446217156
+
+#35: 110-89-4_b.pdf
+difference = 544.2830255599977
